chores:refactor code
